add: add modularity in code
